Removes unnecessary status content_types.
This commit fixes a bug that migration 0002 create_status would add all models instead of only the models that support statuses.  A fix of this migration has been provided in addition to a new migration 0012 to cleanup these statuses for anyone already running the plugin.

In addition to these fixes, I am adding reverse_code for all of the RunPython migrations so that when removing this app, you can do `migrate nautobot_firewall_models zero`.